move getIdJsonUrl, checkPath, startPath and setLangsMenu in another file
